# docs/swagger.yaml
definitions:
  models.Hamburguer:
    properties:
      descricao:
        type: string
      id:
        type: integer
      ingredientes:
        items:
          $ref: '#/definitions/models.HamburguerIngrediente'
        type: array
      preco:
        type: number
    required:
    - descricao
    - preco
    type: object
  models.HamburguerIngrediente:
    properties:
      hamburguerID:
        type: integer
      item:
        $ref: '#/definitions/models.Item'
      itemID:
        type: integer
      quantidade:
        type: integer
    type: object
  models.HamburguerRequest:
    properties:
      descricao:
        type: string
      id:
        type: integer
      ingredientes:
        items:
          $ref: '#/definitions/models.IngredienteRequest'
        minItems: 1
        type: array
      preco:
        type: number
    required:
    - descricao
    - id
    - ingredientes
    - preco
    type: object
  models.HamburguerUpdateRequest:
    properties:
      descricao:
        type: string
      ingredientes:
        items:
          $ref: '#/definitions/models.IngredienteRequest'
        minItems: 1
        type: array
      preco:
        type: number
    required:
    - descricao
    - ingredientes
    - preco
    type: object
  models.IngredienteRequest:
    properties:
      id:
        type: integer
      quantidade:
        minimum: 1
        type: integer
    required:
    - id
    - quantidade
    type: object
  models.Item:
    properties:
      descricao:
        type: string
      extra:
        description: true para "Com açúcar" em bebidas ou "Adicional" em ingredientes
        type: boolean
      id:
        type: integer
      preco:
        type: number
      tipo:
        $ref: '#/definitions/models.TipoItem'
    type: object
...
  models.TipoItem:
    enum:
    - BEBIDA
    - INGREDIENTE
    type: string
    x-enum-varnames:
    - TipoBebida
    - TipoIngrediente
info:
  contact: {}
paths:
  /hamburguers:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os hamburgueres disponíveis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Hamburguer'
            type: array
      summary: Lista todos os hamburgueres
      tags:
      - hamburgueres
    post:
      consumes:
      - application/json
      description: Cria um novo hamburguer com os dados fornecidos
      parameters:
      - description: Dados do Hamburguer
        in: body
        name: hamburguer
        required: true
        schema:
          $ref: '#/definitions/models.HamburguerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Hamburguer'
        "400":
          description: Erro na validação dos dados
          schema:
            type: string
        "404":
          description: Ingrediente não encontrado
          schema:
            type: string
      summary: Cria um novo hamburguer
      tags:
      - hamburgueres
  /hamburguers/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um hamburguer existente
      parameters:
      - description: ID do Hamburguer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Erro ao deletar hamburguer
          schema:
            type: string
        "404":
          description: Hamburguer não encontrado
          schema:
            type: string
      summary: Deleta um hamburguer
      tags:
      - hamburgueres
    get:
      consumes:
      - application/json
      description: Retorna um hamburguer específico baseado no ID
      parameters:
      - description: ID do Hamburguer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hamburguer'
        "404":
          description: Hamburguer não encontrado
          schema:
            type: string
      summary: Busca um hamburguer por ID
      tags:
      - hamburgueres
    put:
      consumes:
      - application/json
      description: Atualiza um hamburguer existente com os dados fornecidos
      parameters:
      - description: ID do Hamburguer
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do Hamburguer
        in: body
        name: hamburguer
        required: true
        schema:
          $ref: '#/definitions/models.HamburguerUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hamburguer'
        "400":
          description: Erro na validação dos dados
          schema:
            type: string
        "404":
          description: Hamburguer não encontrado
          schema:
            type: string
      summary: Atualiza um hamburguer existente
      tags:
      - hamburgueres
  /hamburguers/nome/{nome}:
    get:
      consumes:
      - application/json
      description: Retorna um hamburguer específico baseado no nome
      parameters:
      - description: Nome do Hamburguer
        in: path
        name: nome
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hamburguer'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Busca um hamburguer por nome
      tags:
      - hamburgueres
...
swagger: "2.0"